working product_service. Add grpc, middleware, cors

+++ internal/entity/product.go
@@ -8,7 +8,7 @@ type Product struct {
 	Name        string    `json:"name"`
 	Description string    `json:"description"`
 	Price       float64   `json:"price"`
-	Stock       int       `json:"stock"`
+	Stock       int64     `json:"stock"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
